requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
pytest>=8.3.0
httpx>=0.27.0
